requirements.txt
# 项目依赖包
customtkinter==5.2.2  # 美化版的tkinter界面库
pillow==10.2.0       # 图像处理库
# 在这里添加您的项目所需的包 